+++ README.md
@@ -1,2 +1,2 @@
-# reader-xrd
+# fairmat-readers-xrd
 A repo containg file readers for multiple formats of XRD data.
