+++ src/fairmat_readers_xrd/__init__.py
